Add Week3 Excercise 2: Angular templates
The ng-include won't show anything unless you load it using a local server.
If you can't see the included content in the index.html page, that is because you are facing an XMLHttpRequest problem, your browser cannot load cross-origin requests.

The answer to this problem:

1) You can launch Chrome with --allow-file-access-from-files parameter to disable checks for file://.

2) Change your gulpfile.js:

2.1) In the usemin task change to gulp.src('app/*.html') ...

2.2) Change browserSync.init to:
	browserSync.init(files, {
  server: {
    baseDir: 'dist',
    index: 'index.html'
  },
  reloadDelay: 1000
});

Yoy can find this solution in Week3 forum.